requirements.txt
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-qdrant>=0.2.0
qdrant-client>=1.12.0
python-dotenv>=1.0
fastapi
uvicorn[standard]
requests>=2.31.0
